layout: rename digital_pnr to pnr and declare its API with @public
The package moves to ordec.layout.pnr, with tests/test_pnr.py and
docs/ref/pnr.rst following. Public members are now declared with @public
at their definitions in flow.py and route.py, and the package __init__ is
two star imports like ordec.core. is_extlibrary_leaf becomes public as
the documented is_leaf default. The docs paragraph that justified the
digital prefix now states the standard-cell requirement without it.
