+++ tests/lib/pnr_cells.ord
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 """
-Synthetic cells for the place-and-route tests (tests/test_digital_pnr.py,
+Synthetic cells for the place-and-route tests (tests/test_pnr.py,
 tests/test_ihp130_pnr.py).
 
 Each has a circuit shape the others do not, so a clean result is evidence the
@@ -32,7 +32,7 @@ from ordec.core import *
 from ordec.extlibrary import ExtLibrary
 from ordec.lib import ihp130
 from ordec.lib.ihp130 import SG13G2, Nmos, lef_pin_rects, sg13g2_grid
-from ordec.layout.digital_pnr import place_and_route
+from ordec.layout.pnr import place_and_route
 
 stdcells = ExtLibrary()
 stdcells.read_lef(ihp130.pdk().stdcell_lef)
